# gomaxprocs: Use a ticker (30s) to set gomaxprocs