--- a/beater/beater.go
+++ b/beater/beater.go
@@ -96,22 +96,11 @@ func NewCreator(args CreatorParams) beat.Creator {
 			libbeatMonitoringRegistry: monitoring.Default.GetRegistry("libbeat"),
 		}
 
-		// Use `maxprocs` to change the GOMAXPROCS respecting any CFS quotas, if
-		// set. This is necessary since the Go runtime will default to the number
-		// of CPUs available in the  machine it's running in, however, when running
-		// in a container or in a cgroup with resource limits, the disparity can be
-		// extreme.
-		// Having a significantly greater GOMAXPROCS set than the granted CFS quota
-		// results in a significant amount of time spent "throttling", essentially
-		// pausing the the running OS threads for the throttled period.
-		_, err := maxprocs.Set(maxprocs.Logger(logger.Infof))
-		if err != nil {
-			logger.Errorf("failed to set GOMAXPROCS: %v", err)
-		}
 		var elasticsearchOutputConfig *agentconfig.C
 		if hasElasticsearchOutput(b) {
 			elasticsearchOutputConfig = b.Config.Output.Config()
 		}
+		var err error
 		bt.config, err = config.NewConfig(bt.rawConfig, elasticsearchOutputConfig)
 		if err != nil {
 			return nil, err
@@ -164,6 +153,18 @@ func (bt *beater) Run(b *beat.Beat) error {
 }
 
 func (bt *beater) run(ctx context.Context, cancelContext context.CancelFunc, b *beat.Beat) error {
+	// Use `maxprocs` to change the GOMAXPROCS respecting any CFS quotas, if
+	// set. This is necessary since the Go runtime will default to the number
+	// of CPUs available in the  machine it's running in, however, when running
+	// in a container or in a cgroup with resource limits, the disparity can be
+	// extreme.
+	// Having a significantly greater GOMAXPROCS set than the granted CFS quota
+	// results in a significant amount of time spent "throttling", essentially
+	// pausing the the running OS threads for the throttled period.
+	// Since the quotas may be updated without restarting the process, the
+	// GOMAXPROCS are adjusted every 30s.
+	go adjustMaxProcs(ctx, 30*time.Second, bt.logger)
+
 	tracer, tracerServer, err := initTracing(b, bt.config, bt.logger)
 	if err != nil {
 		return err
@@ -1025,4 +1026,24 @@ func queryClusterUUID(ctx context.Context, esClient elasticsearch.Client) error 
 
 	s.Set(response.ClusterUUID)
 	return nil
+}
+
+func adjustMaxProcs(ctx context.Context, d time.Duration, logger *logp.Logger) error {
+	setMaxProcs := func() {
+		if _, err := maxprocs.Set(maxprocs.Logger(logger.Infof)); err != nil {
+			logger.Errorf("failed to set GOMAXPROCS: %v", err)
+		}
+	}
+	// set the gomaxprocs immediately.
+	setMaxProcs()
+	ticker := time.NewTicker(d)
+	defer ticker.Stop()
+	for {
+		select {
+		case <-ctx.Done():
+			return ctx.Err()
+		case <-ticker.C:
+			setMaxProcs()
+		}
+	}
 }